Finally got script to work on NFS shares running Linux (Ubuntu)

+++ DiskConstrictorOO.py
@@ -1,9 +1,9 @@
 # File: TestProject.py
 """ This script will allow an unlimited number of write/read/compare threads to a network volume.
 The file size will vary from 37 bytes to 111 million bytes. If you only run one thread, you will always
-get the largest possible size (111 million). If you only run two threads, you will always get the largest
-and smallest possible sizes (37 bytes). More than three threads will generate a random file size between
-37 and 111 million, exclusive.
+get the largest possible size (111 MB). If you only run two threads, you will always get the largest
+and smallest possible sizes (111 MB & 37 bytes). More than three threads will generate a random file size that
+is a multiple of 37 and is between 37 and 111 million, exclusive.
 You can also use this script as a crude benchmark program by testing only a single thread and setting
 the "gShowXferSpeeds" to True
 """
@@ -23,11 +23,12 @@ gShowXferSpeeds     =   False
 gMaxFiles           =   10000
 gOriginalDir        =   " "
 gOKToStartThreads   =   False
+gInjectError        =   False
 
 #---------------------- These are basically just constants -----------------
 gOneMegabyte        =   1000000
 gMaxXFerSize        =   gOneMegabyte * 3    # max out at a 111 MB
-#gMaxXFerSize        =   10000    # max out at a 111 MB
+#gMaxXFerSize        =   1000    # max out at a 111 MB
 gTotalUniqueChars   =   37    #there are 26 characters in alphabet + 0-9 + carriage return = 37
 
 # noinspection PyPep8Naming,PyPep8Naming
@@ -46,7 +47,7 @@ class IOTester:
         elif instanceNumber == 1:
             self.charStrMultiple    =   1
         else:
-            self.charStrMultiple    =   random.randrange(2,gMaxXFerSize, 1) # pick a number between 1 and max size only
+            self.charStrMultiple    =   random.randrange(2, gMaxXFerSize, 1) # pick a number between 1 and max size only
         self.sourceBuffer           =   bytes(b'ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789\n' * self.charStrMultiple)
         self.TotalBytes             =   gTotalUniqueChars * self.charStrMultiple
         self.megsXferred            =   self.TotalBytes / gOneMegabyte
@@ -68,10 +69,10 @@ class IOTester:
         if gDebugLevel > 0:
             print("\nCreating New Test File...\n")
         for j in range(gMaxFiles):
-            if not os.path.isfile(self.testFileName):            #is there no file with this name already?
-                break                                               #yep, break out of loop
+            if os.path.isfile(self.testFileName):            #is there file with this name already?
+                self.testFileName    =   "testfile{0}.txt".format(j+2)   #yes, try next file name
             else:
-                self.testFileName    =   "testfile{0}.txt".format(j+2)   #nope, try next file name
+                break                                               #no, break out of loop
         # noinspection PyUnboundLocalVariable
         if j == gMaxFiles - 1:
             print("Sorry, exceeded the number of unique file names (10,000)\nExiting program. Try deleting all the test files in the 'TestFiles' directory")
@@ -80,7 +81,10 @@ class IOTester:
             if gDebugLevel > 0:
                 print("Found a test file name for Instance:", self.instanceNo, "The name is: ", self.testFileName)
                 print("Thread from instance", self.instanceNo, "starting a new test cycle")
-            self.myFileH = open(self.testFileName, "wb+", buffering=0)   # this is good enough for SMB on MacOS
+
+            if sys.platform != "linux":
+                self.myFileH = open(self.testFileName, mode="wb+", buffering=0)   # this is good enough for SMB on MacOS
+
             if sys.platform == "darwin":                      # Disable caching on Mac/afp
                 # noinspection PyUnusedLocal
                 ignoreResult    =   fcntl.fcntl(self.myFileH, fcntl.F_NOCACHE, 1)
@@ -91,10 +95,22 @@ class IOTester:
         return
 
     def CompareWholeFile(self):
+        global gDebugLevel
         global gkeyboardinputstr
         global gOriginalDir
+        global gInjectError
         try:
-            xFerBytes    =   self.myFileH.readinto(self.destBuffer)
+            if sys.platform == "linux": #under Linux we assume we are running NFS. You cannot disable client side caching so we do the open/close/delete sequence to defeat the cache
+                os.utime(self.testFileName)
+                self.myFileH.seek(0, io.SEEK_SET)
+                xFerBytes       =   self.myFileH.readinto(self.destBuffer)
+                self.myFileH.close()
+                os.remove(os.path.realpath(self.testFileName))
+            else:
+                self.myFileH.seek(0, io.SEEK_SET)
+                xFerBytes    =   self.myFileH.readinto(self.destBuffer)
+            if gInjectError:
+                self.destBuffer[15]   =   65;# Put an "A" where the "P" should be in the first line to force a mis-compare error
             if xFerBytes != self.TotalBytes:
                 gkeyboardinputstr = 'p'  # Pause all tests
                 print("Instance ", self.instanceNo, "did not get the expected number of bytes. Pausing all tests.\n")
@@ -122,14 +138,18 @@ class IOTester:
         global gOKToStartThreads
         global gDebugLevel
         global gShowXferSpeeds
+        global gkeyboardinputstr
 
         while not gOKToStartThreads:   # Wait to start testing thread until all test class instances have been initialized
-            time.sleep(.5)
             if gDebugLevel > 0:
                 print("Entering thread loop for instance:", self.instanceNo)
         while True:
             if CheckForNewKeyboardInput():
                 break
+            if sys.platform == "linux":
+                self.myFileH    =   open(self.testFileName, mode="wb+", buffering=0)
+            else:
+                self.myFileH.seek(0, io.SEEK_SET)
             if gDebugLevel > 0:
                 print("Starting Write...")
             if gShowXferSpeeds:
@@ -139,7 +159,8 @@ class IOTester:
             else:
                 if gDebugLevel > 1:
                     print("sourceBuffer from instance ", self.instanceNo, " is: ", len(self.sourceBuffer), " bytes in size")
-# Ugly hack: Not sure why, but Windows needs this to be inside a timer function or the threads will eventually crash. Maybe i/o is not thread safe in Windows?
+# Ugly hack: Not sure why, but Windows needs this to be inside a timer function or the threads will eventually crash.
+# Maybe i/o is not thread safe in Windows? Root cause should be investigated and a better solution found.
                 if sys.platform == "win32":
                     t           =   timeit.Timer(self.WriteTestPattern)
                     totalTime   =   t.timeit(number=1)
@@ -147,15 +168,12 @@ class IOTester:
                     try:
                         self.myFileH.write(self.sourceBuffer)
                     except:
-                        print("Pausing Tests. Write file failed. Error: ", sys.exc_info()[0])
+                        print("Pausing Tests. Write Error: ", sys.exc_info()[0])
                         gkeyboardinputstr = "p"
 
             if CheckForNewKeyboardInput():
                 break
-
-            if gDebugLevel > 0:
-                print("Starting Seek...")
-            self.myFileH.seek(0, io.SEEK_SET)
+            
             if gDebugLevel > 0:
                 print("Starting Read/Compare...")
             if gShowXferSpeeds:
@@ -165,12 +183,12 @@ class IOTester:
             else:
                 self.CompareWholeFile()
 
-            if gDebugLevel > 0:
-                print("Starting Seek...")
-            self.myFileH.seek(0, io.SEEK_SET)
+            if CheckForNewKeyboardInput():  #if true, user wants to quit. If pause, call will block until user quits or presses 'r'
+                break
 
         self.myFileH.close()
-        os.remove(os.path.realpath(self.testFileName))
+        if os.path.exists(os.path.realpath(self.testFileName)):
+            os.remove(os.path.realpath(self.testFileName))
         self.threadTerminated   =   True
         return
 
@@ -192,7 +210,7 @@ def getkeyboardinput_thread():
         if gkeyboardinputstr == "q":   # if the user wants to quit, we need to kill this thread
             break
         elif gkeyboardinputstr == "p":
-            print("\nAll tests paused. Press 'r' to resume or 'q' to quit")
+            print("\nAll tests paused. Press 'r' to resume or 'q' to quit\n")
         elif gkeyboardinputstr == "r":
             print("\nResuming tests")
     return
@@ -205,6 +223,7 @@ def setTestWorkingDirectory():  #need to return actual error in future version
             print("\nPlease make sure that you have only the public share of the test drive (UUT) mounted. You can mount it using whatever protocol you wish")
             return 1
     elif sys.platform == "win32":
+        print("\nPlease make sure you have turned OpLocks off on the UUT!!!!!!!!!!!!!!!")
         try:
             myStr    =   input("Please <enter> the IP address of the test drive (UUT): ")
             os.chdir("\\\\" + myStr + "\\Public\\")
@@ -212,18 +231,15 @@ def setTestWorkingDirectory():  #need to return actual error in future version
             print("\nPlease make sure that you have the public share of the test drive (UUT) mounted and that you have entered the correct IP address")
             return 1
     elif sys.platform == "linux":
-        print("\nTo run this program under Linux, you must make sure to run python as root")
-        print("e.g. 'sudo python3.5 DiskConstrictorOO.py'")
-        print("\nBefore you run this script, you must create a local mountpoint at '/mnt/Constrictor'")
+        print("\nBefore you start the tests, you must create a local mountpoint at '/mnt/Constrictor'")
         print("e.g. 'sudo mkdir /mnt/Constrictor'")
-        print("\nNext, you must mount the public share using the desired protocol")
-        print("e.g. for NFS use something like: 'sudo mount 192.168.1.137:/nfs/Public' /mnt/Constrictor")
-        print("e.g. for SMB use something like: 'sudo mount //192.168.1 137/Public /mnt/Constrictor")
-        print("Now you are ready to run the script as root")
+        print("\nNext, you must mount the public share to the local mountpoint using the desired protocol and disable attribute caching for NFS Volumes")
+        print("For NFS use something like: 'sudo mount -o noac 192.168.1.137:/nfs/Public' /mnt/Constrictor")
+        print("For SMB use something like: 'sudo mount //192.168.1 137/Public /mnt/Constrictor")
         try:
             os.chdir("/mnt/Constrictor")
         except:
-            print("Unable to change active directory to /mnt/Constrictor.\nPlease make sure the directory exists and that Python is running as root.")
+            print("Unable to change active directory to /mnt/Constrictor.\nPlease make sure the directory exists.")
             return 1
     else:
         print("Current reported platform is:", sys.platform)
@@ -262,7 +278,7 @@ def main():
 
     gkeyboardinputstr    =   "A"
     print("\nTo Pause Press:  <p> <Enter>\nTo Resume Press: <r> <Enter>\nTo Quit Press:   <q> <Enter>\n")
-    testThreadCount     =   eval(input("\nFor best performance, you need about of 111 MB of free memory per thread\nPlease enter the number of test threads you want to use:"))
+    testThreadCount     =   eval(input("\nFor best performance, you need an average of about of 111 MB of free memory per thread\nPlease enter the number of test threads you want to use:"))
     print("There will be", testThreadCount, "test thread(s) created")
     kbThread            =  threading.Thread(target=getkeyboardinput_thread)
     kbThread.start()
@@ -274,7 +290,7 @@ def main():
     # noinspection PyRedeclaration,PyRedeclaration
     gOKToStartThreads   =   True
 
-    print("\nRunning Test(s). Start time:", time.asctime( time.localtime(time.time()) ))
+    print("\nRunning Test(s).\nStart time:", time.asctime( time.localtime(time.time()) ))
 
 # this would be a good place to put an animation in the terminal window so
 # that people can see that the program is still alive
@@ -286,8 +302,5 @@ def main():
         while newTester[i].threadTerminated == False:
             time.sleep(.1)
     print("\nEnd time:", time.asctime( time.localtime(time.time()) ))
-
-    #if sys.platform == "linux":
-        #os.system("sudo umount /mnt/Crusher")
 
 main()
